requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
et-xmlfile==1.1.0
idna==3.4
mysql-connector==2.2.9
mysql-connector-repackaged==0.3.1
numpy==1.25.1
openpyxl==3.1.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
urllib3==2.0.3
watchdog==3.0.0
zeroc-ice==3.7.9.1
